Split metadata in multiple files
